fix: moved vidazoo buildRequestData into lib
